requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.0.0
img2pdf==0.4.3
lxml==4.8.0
packaging==21.3
pikepdf==5.0.1
Pillow==9.0.1
pyparsing==3.0.7
sqlparse==0.4.2
whitenoise==6.0.0
